requirement.txt
boto==2.49.0
boto3==1.9.169
botocore==1.12.169
certifi==2019.3.9
chardet==3.0.4
Click==7.0
docutils==0.14
Flask==1.0.3
gensim==3.7.3
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.9.4
joblib==0.13.2
MarkupSafe==1.1.1
numpy==1.16.4
pandas==0.24.2
pid==2.2.3
python-dateutil==2.8.0
python-Levenshtein==0.12.0
pytz==2019.1
requests==2.22.0
s3transfer==0.2.1
scikit-learn==0.21.2
scipy==1.3.0
six==1.12.0
sklearn==0.0
smart-open==1.8.4
tqdm==4.32.1
urllib3==1.25.3
Werkzeug==0.15.4
